src/main/python/services/dss_images/requirements.txt
Flask==3.0.0
google-cloud-datastore
numpy
onnxruntime
opencv-python-headless
pillow